move network to it's owm file
