+++ source/type/network.h
@@ -1,0 +1,61 @@
+/* Copyright © 2001-2014, Canal TP and/or its affiliates. All rights reserved.
+
+This file is part of Navitia,
+    the software to build cool stuff with public transport.
+
+Hope you'll enjoy and contribute to this project,
+    powered by Canal TP (www.canaltp.fr).
+Help us simplify mobility and open public transport:
+    a non ending quest to the responsive locomotion way of traveling!
+
+LICENCE: This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+Stay tuned using
+twitter @navitia
+IRC #navitia on freenode
+https://groups.google.com/d/forum/navitia
+www.navitia.io
+*/
+
+#pragma once
+#include "type/type_interfaces.h"
+#include "type/fwd_type.h"
+#include <set>
+
+namespace navitia {
+namespace type {
+
+struct Network : public Header, HasMessages {
+    std::string name;
+    const static Type_e type = Type_e::Network;
+    std::string address_name;
+    std::string address_number;
+    std::string address_type_name;
+    std::string phone_number;
+    std::string mail;
+    std::string website;
+    std::string fax;
+    int sort = std::numeric_limits<int>::max();
+
+    std::vector<Line*> line_list;
+    std::set<Dataset*> dataset_list;
+
+    template <class Archive>
+    void serialize(Archive& ar, const unsigned int);
+
+    Indexes get(Type_e type, const PT_Data& data) const;
+    bool operator<(const Network& other) const;
+};
+}  // namespace type
+}  // namespace navitia
